deduplicated some of the results
if there were two of the same number in the number input, they
used to be treated as unique. now if two leafs are the same, they
are treated as if they were the same.

Moreover, since addition and multiplication are commutative, if
two operands of those operation are just mirrored, it also gets
deduplicated.