Deduplicated clones caused by permutation of same kind operation
so let's say you have 4 numbers multiplied together

2 3 4 5
it used to be that order of operations such as

2 x 3 = 6
4 x 5 = 20
20 x 6 = 120

was considered different than

2 x 4 = 8
3 x 5 = 15
15 x 8 = 120

and to some, it might be considered different but they are all just

2 x 3 x 4 x 5

the order of which of these numbers you multiply by shouldn't matter
and shouldn't be considered a different solution. same goes for
addition, subtraction and division as long as order of operations
are followed.

this also improves performance.